Añadido nuevo sistema doble de buffers diferenciado para la creacion de los ficheros intermedios y para la lectura de los mismos. Se ha lanzado un ejemplo de un metagenoma con 678.471 Reads que generan 296.381.375 palabras. El programa generador de diccionarios ha tardado en un portatil personal los siguientes tiempos:    > 7min 34s en generar los ficheros intermedios y liberar el buffer de escritura.    > 2min 40s en escribir las primeras 100millones de palabras (Velocidad: 62.500 palabras minuto).    > 3min 26s en escribir las segundas 100millones de palabras (Velocidad: 55.556 palabras minuto). Despues de esto el ejemplo CRASHEA y no termina con las ultimas 96millones de palabras. Nota: se sabe que el programa termina con un ejemplo pequeño. Se testeara la razon del error. Tiempo del programa antes del crasheo: 16min 42s. Tiempo entre la ultima escritura conocida y el crasheo: 3min 27s. Nota: se desconoce si las ultimas 96millones se llegan a escribir o no (mensaje de testeo no añadido en el codigo).

+++ HSPandCSB/src/dictFun.c
@@ -245,7 +245,7 @@ uint64_t lowestWord(wentry *words,uint64_t length){
  *  @param sameThanLastWord boolean value that indicate if the current word is the same than the last written.
  *  @param words equal than last written.
  */
-inline void writeWord(wentry *word, FILE* w, FILE* p, bool sameThanLastWord, uint16_t *words){
+inline void writeWord(wentry *word, FILE* w, FILE* p, bool sameThanLastWord, uint64_t *words){
 	uint64_t aux;
 	if(!sameThanLastWord){ // Write new word
 		fwrite(words,sizeof(uint16_t),1,w); // Write num of repetitions
